Kesme Diagramı Eklendi

+++ cisimmodeli.h
@@ -14,33 +14,33 @@ public:
     QString tipIsmiAl() { return tipIsmim[_tip]; }
     int tipAl() const { return _tip; }
     int tipAl(QString tipIsmi) { return tipIsmim.key(tipIsmi); }
-    int noktaKonumuAl() const { return _noktaKonumu; }
-    int noktaKuvvetiAl() const { return _noktaKuvveti; }
-    int baslangciKonumuAl() const { return _baslangicKonumu; }
-    int bitisKonumuAl() const { return _bitisKonumu; }
-    int baslangicKuvvetiAl() const { return _baslangicKuvveti; }
-    int bitisKuvvetiAl() const { return _bitisKuvveti; }
-    int momentAl() const { return _moment; }
+    double noktaKonumuAl() const { return _noktaKonumu; }
+    double noktaKuvvetiAl() const { return _noktaKuvveti; }
+    double baslangciKonumuAl() const { return _baslangicKonumu; }
+    double bitisKonumuAl() const { return _bitisKonumu; }
+    double baslangicKuvvetiAl() const { return _baslangicKuvveti; }
+    double bitisKuvvetiAl() const { return _bitisKuvveti; }
+    double momentAl() const { return _moment; }
 
     void tipIsmiAta(QString tipIsmi) { _tipIsmi = tipIsmi; }
     void tipAta(int tip) { _tip = tip; }
-    void noktaKonumuAta(int noktaKonumu) { _noktaKonumu = noktaKonumu; }
-    void noktaKuvvetiAta(int noktaKuvveti ) { _noktaKuvveti = noktaKuvveti; }
-    void baslangciKonumuAta(int baslangicKonumu ) { _baslangicKonumu = baslangicKonumu; }
-    void bitisKonumuAta(int bitisKonumu ) { _bitisKonumu = bitisKonumu; }
-    void baslangicKuvvetiAta(int baslangicKuvveti ) { _baslangicKuvveti = baslangicKuvveti; }
-    void bitisKuvvetiAta(int bitisKuvveti ) { _bitisKuvveti = bitisKuvveti; }
-    void momentAta(int moment) { _moment = moment; }
+    void noktaKonumuAta(double noktaKonumu) { _noktaKonumu = noktaKonumu; }
+    void noktaKuvvetiAta(double noktaKuvveti ) { _noktaKuvveti = noktaKuvveti; }
+    void baslangciKonumuAta(double baslangicKonumu ) { _baslangicKonumu = baslangicKonumu; }
+    void bitisKonumuAta(double bitisKonumu ) { _bitisKonumu = bitisKonumu; }
+    void baslangicKuvvetiAta(double baslangicKuvveti ) { _baslangicKuvveti = baslangicKuvveti; }
+    void bitisKuvvetiAta(double bitisKuvveti ) { _bitisKuvveti = bitisKuvveti; }
+    void momentAta(double moment) { _moment = moment; }
 private:
     QString _tipIsmi;
     int _tip;
-    int _noktaKonumu;
-    int _noktaKuvveti;
-    int _baslangicKonumu;
-    int _bitisKonumu;
-    int _baslangicKuvveti;
-    int _bitisKuvveti;
-    int _moment;
+    double _noktaKonumu;
+    double _noktaKuvveti;
+    double _baslangicKonumu;
+    double _bitisKonumu;
+    double _baslangicKuvveti;
+    double _bitisKuvveti;
+    double _moment;
     QMap<int,QString> tipIsmim;
 };
